ref(node): Export encodeStr , hexTable and isHexTable
