+++ node/querystring.ts
@@ -1,6 +1,6 @@
 // Copyright 2018-2021 the Deno authors. All rights reserved. MIT license.
 import { Buffer } from "./buffer.ts";
-import { ERR_INVALID_URI } from "./_errors.ts";
+import { encodeStr, hexTable } from "./internal/querystring.ts";
 
 /**
  * Alias of querystring.parse()
@@ -13,69 +13,6 @@ export const decode = parse;
  * @legacy
  */
 export const encode = stringify;
-
-const hexTable = new Array(256);
-for (let i = 0; i < 256; ++i) {
-  hexTable[i] = "%" + ((i < 16 ? "0" : "") + i.toString(16)).toUpperCase();
-}
-
-function encodeStr(
-  str: string,
-  noEscapeTable: Int8Array,
-  hexTable: string[],
-): string {
-  const len = str.length;
-  if (len === 0) return "";
-
-  let out = "";
-  let lastPos = 0;
-
-  for (let i = 0; i < len; i++) {
-    let c = str.charCodeAt(i);
-    // ASCII
-    if (c < 0x80) {
-      if (noEscapeTable[c] === 1) continue;
-      if (lastPos < i) out += str.slice(lastPos, i);
-      lastPos = i + 1;
-      out += hexTable[c];
-      continue;
-    }
-
-    if (lastPos < i) out += str.slice(lastPos, i);
-
-    // Multi-byte characters ...
-    if (c < 0x800) {
-      lastPos = i + 1;
-      out += hexTable[0xc0 | (c >> 6)] + hexTable[0x80 | (c & 0x3f)];
-      continue;
-    }
-    if (c < 0xd800 || c >= 0xe000) {
-      lastPos = i + 1;
-      out += hexTable[0xe0 | (c >> 12)] + hexTable[0x80 | ((c >> 6) & 0x3f)] +
-        hexTable[0x80 | (c & 0x3f)];
-      continue;
-    }
-    // Surrogate pair
-    ++i;
-
-    // This branch should never happen because all URLSearchParams entries
-    // should already be converted to USVString. But, included for
-    // completion's sake anyway.
-    if (i >= len) throw new ERR_INVALID_URI();
-
-    const c2 = str.charCodeAt(i) & 0x3ff;
-
-    lastPos = i + 1;
-    c = 0x10000 + (((c & 0x3ff) << 10) | c2);
-    out += hexTable[0xf0 | (c >> 18)] +
-      hexTable[0x80 | ((c >> 12) & 0x3f)] +
-      hexTable[0x80 | ((c >> 6) & 0x3f)] +
-      hexTable[0x80 | (c & 0x3f)];
-  }
-  if (lastPos === 0) return str;
-  if (lastPos < len) return out + str.slice(lastPos);
-  return out;
-}
 
 /**
  * replaces encodeURIComponent()
